requirements.txt
discord
asyncio
python-dotenv
aiohttp
os